chore: move tokenBucket to Algorithm namespace
